requirements.txt
websocket-client
pandas
numpy
scikit-learn
